feat: fixed int32 int64 car year mismatch

+++ api/car.go
@@ -12,7 +12,7 @@ type createCarRequest struct {
 	OwnerID int64  `json:"owner_id"`
 	Make    string `json:"make"`
 	Model   string `json:"model"`
-	Year    int64  `json:"year"`
+	Year    int32  `json:"year"`
 }
 
 func (server *Server) createCar(ctx *gin.Context) {
